Added verapp_profile_check_source_files_conformance to actuallly run vera++ on the source files that build a particular target.
vera++ has a curious behaviour of not returning an error code when it finds
problems (even though it prints to the stderr) so we also provide a wrapper
to run vera++ through. If that wrapper finds anything on the stderr it will
force a non-zero return code.

Added LICENCE.md file